requirements.txt
Flask
flask_sqlalchemy
python-dotenv
Gunicorn
psycopg2